+++ src/lib.rs
@@ -18,6 +18,7 @@ mod debugfs;
 mod error;
 mod maps;
 mod perf;
+mod perf_map_poller;
 
 pub use blueprint::{ProgramBlueprint, SectionType};
 pub use error::OxidebpfError;
@@ -30,13 +31,13 @@ use bpf::{
     BpfAttr, MapConfig, SizedBpfAttr,
 };
 use debugfs::{get_debugfs_mount_point, mount_debugfs_if_missing};
-
-use maps::{PerCpu, PerfEvent, PerfMap, ProgMap};
+use maps::{PerCpu, PerfMap, ProgMap};
 use perf::{
     constant::{perf_event_sample_format, perf_sw_ids, perf_type_id},
     syscall::{attach_kprobe, attach_kprobe_debugfs, attach_uprobe, attach_uprobe_debugfs},
     PerfEventAttr, PerfSample, PerfWakeup,
 };
+use perf_map_poller::perf_map_poller;
 
 use std::{
     collections::{HashMap, HashSet},
@@ -45,16 +46,13 @@ use std::{
     io::{BufRead, BufReader, BufWriter, Write},
     os::unix::io::RawFd,
     sync::{Arc, Condvar, Mutex},
-    thread,
     time::Duration,
 };
 
 use crossbeam_channel::{bounded, Receiver, Sender};
 use lazy_static::lazy_static;
 use libc::{c_int, pid_t};
-use mio::{unix::SourceFd, Events, Interest, Poll, Token};
-use nix::errno::Errno;
-use slog::{crit, error, info, o, Logger};
+use slog::{crit, info, o, Logger};
 use slog_atomic::{AtomicSwitch, AtomicSwitchCtrl};
 
 lazy_static! {
@@ -89,17 +87,25 @@ pub struct CapUserData {
     inheritable: u32,
 }
 
-/// Message format for messages sent back across the channel. It includes
-/// the map name, cpu id, and message data.
+/// Message format for messages sent back across the channel
+///
+/// Messages could either be dropped in case of messages lost, usually
+/// caused by a a filled up perf buffer, or an event with source and
+/// data
 #[derive(Debug)]
-pub struct PerfChannelMessage(
-    /// The name of the map this message is from.
-    pub String,
-    /// The cpuid of the CPU this message is from.
-    pub i32,
-    /// The data in the message.
-    pub Vec<u8>,
-);
+pub enum PerfChannelMessage {
+    /// A count of how many messages were lost
+    Dropped(u64),
+    /// A received message with information about its source
+    Event {
+        /// The name of the map this message is from.
+        map_name: String,
+        /// The cpuid of the CPU this message is from.
+        cpuid: i32,
+        /// The data in the message.
+        data: Vec<u8>,
+    },
+}
 
 #[derive(Clone)]
 struct Channel {
@@ -631,7 +637,7 @@ impl<'a> ProgramGroup<'a> {
             bounded(event_buffer_size);
         let channel = Channel { tx, rx };
         ProgramGroup {
-            event_buffer_size,
+            event_buffer_size: event_buffer_size,
             channel,
             loaded_version: None,
             mem_limit: None,
@@ -1309,174 +1315,6 @@ impl<'a> Drop for ProgramVersion<'a> {
             drop_debugfs_uprobes(debugfs_mount);
             drop_debugfs_kprobes(debugfs_mount);
         }
-    }
-}
-
-fn perf_map_poller(
-    perfmaps: Vec<PerfMap>,
-    tx: Sender<PerfChannelMessage>,
-    polling_delay: Duration,
-    polling_policy: SchedulingPolicy,
-    polling_signal: Arc<(Mutex<bool>, Condvar)>,
-) {
-    let native_id = match polling_policy {
-        SchedulingPolicy::Deadline(_, _, _) => {
-            // SAFETY: this syscall is always successful
-            unsafe { libc::syscall(libc::SYS_gettid) as libc::pthread_t }
-        }
-        _ => thread_priority::thread_native_id(),
-    };
-    let priority = polling_policy.into();
-    let policy = polling_policy.into();
-
-    // This call throws errors if the passed in priority and policies don't match, so we need
-    // to ensure that it's what's expected (1 to 99 inclusive for realtime, set of 3 nanosecond
-    // counts for realtime deadline, 0 for all others).
-    if let Err(e) = thread_priority::set_thread_priority_and_policy(native_id, priority, policy) {
-        error!(
-            LOGGER.0,
-            "perf_map_poller(); could not set thread priority, continuing at inherited: {:?}", e
-        );
-    };
-
-    // Once we've set our scheduling policy and priority, we'll want to set the niceness value
-    // (if relevant).
-    match polling_policy {
-        SchedulingPolicy::Other(polling_priority) | SchedulingPolicy::Batch(polling_priority) => {
-            // SAFETY: continuing at the default is not fatal, casting i8 to i32 is safe, clamp
-            unsafe {
-                let polling_priority = polling_priority.clamp(-20, 19);
-                if libc::nice(polling_priority as i32) < 0 {
-                    let errno = nix::errno::Errno::from_i32(nix::errno::errno());
-                    error!(
-                        LOGGER.0,
-                        "perf_map_poller(); could not set niceness, continuing at 0: {:?}", errno
-                    );
-                }
-            };
-        }
-        // we don't need to set a niceness value for anything else
-        _ => {}
-    }
-
-    let mut poll = match Poll::new() {
-        Ok(p) => p,
-        Err(e) => {
-            crit!(
-                LOGGER.0,
-                "perf_map_poller(); error creating poller: {:?}",
-                e
-            );
-            return;
-        }
-    };
-
-    let mut tokens: HashMap<Token, PerfMap> = HashMap::new();
-    for p in perfmaps {
-        let token = Token(p.ev_fd as usize);
-
-        if let Err(e) = poll
-            .registry()
-            .register(&mut SourceFd(&p.ev_fd), token, Interest::READABLE)
-        {
-            crit!(
-                LOGGER.0,
-                "perf_map_poller(); error registering poller: {:?}",
-                e
-            );
-            return;
-        }
-
-        tokens.insert(token, p);
-    }
-
-    {
-        // now that the perfmap fd's are registered, we can signal to the main thread that
-        // event polling is ready.
-        let (lock, cvar) = &*polling_signal;
-        match lock.lock() {
-            Ok(mut polling_is_ready) => {
-                *polling_is_ready = true;
-                cvar.notify_one();
-            }
-            Err(e) => {
-                info!(
-                    LOGGER.0,
-                    "perf_map_poller(); error grabbing cond mutex: {:?}", e
-                );
-            }
-        }
-    }
-
-    let mut events = Events::with_capacity(1024);
-
-    // for tracking dropped event statistics inside the loop
-    let mut dropped = 0_i32;
-    let mut processed = 0_i32;
-
-    'outer: loop {
-        match poll.poll(&mut events, Some(Duration::from_millis(100))) {
-            Ok(_) => {}
-            Err(e) => match nix::errno::Errno::from_i32(nix::errno::errno()) {
-                Errno::EINTR => continue,
-                _ => {
-                    crit!(
-                        LOGGER.0,
-                        "perf_map_poller(); unrecoverable polling error: {:?}",
-                        e
-                    );
-                    return;
-                }
-            },
-        }
-        let mut perf_events: Vec<(String, i32, Option<PerfEvent>)> = Vec::new();
-        events
-            .iter()
-            .filter(|event| event.is_readable())
-            .filter_map(|e| tokens.get(&e.token()))
-            .for_each(|perfmap| loop {
-                match perfmap.read() {
-                    Ok(perf_event) => {
-                        perf_events.push((
-                            perfmap.name.to_string(),
-                            perfmap.cpuid() as i32,
-                            perf_event,
-                        ));
-                    }
-                    Err(OxidebpfError::NoPerfData) => {
-                        // we're done reading
-                        return;
-                    }
-                    Err(e) => {
-                        crit!(LOGGER.0, "perf_map_poller(); perfmap read error: {:?}", e);
-                        return;
-                    }
-                }
-            });
-        for event in perf_events.into_iter() {
-            let message = match event.2 {
-                None => continue,
-                Some(PerfEvent::Lost(l)) => {
-                    dropped = dropped.wrapping_add(l.count as i32);
-                    if (dropped >= 1000 || processed >= 10000) && dropped > 0 {
-                        let d = dropped;
-                        dropped = 0;
-                        processed = 0;
-                        PerfChannelMessage("DROPPED".to_owned(), d, vec![])
-                    } else {
-                        continue;
-                    }
-                }
-                Some(PerfEvent::Sample(e)) => PerfChannelMessage(event.0, event.1, e.data),
-            };
-            match tx.send(message) {
-                Ok(_) => {}
-                Err(_) => break 'outer,
-            };
-
-            processed = processed.wrapping_add(1);
-        }
-        thread::sleep(polling_delay);
     }
 }
